Adjustments to logic based on feedback from CF
Prior assumption was that with regards to the number of chargeable days,
the checkout date was included in that amount, while the checkin day was
not.  After feedback from CF, it appears I had it backwards.

I adjusted the chargeable-days computing logic to skip the checkout
date, but include the checkin date.

One unit test needed adjustment with this new info.  I also updated
readme file to reflect this new info.